feat: parse delete product exception

+++ src/main/java/com/flamarion/productms/services/product/DeleteProductService.java
@@ -1,8 +1,13 @@
 package com.flamarion.productms.services.product;
 
+import com.flamarion.productms.core.http.exceptions.HttpException;
+import com.flamarion.productms.entities.Product;
 import com.flamarion.productms.repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class DeleteProductService {
@@ -13,6 +18,10 @@ public class DeleteProductService {
         this.productRepository = productRepository;
     }
     public void execute(String id) {
+        Optional<Product> productToDelete = this.productRepository.findById(id);
+
+        if (productToDelete.isEmpty()) throw new HttpException(HttpStatus.NOT_FOUND, "Produto não encontrado");
+
         this.productRepository.deleteById(id);
     }
 }
